Fixing Bugs

+++ lib/models/section.dart
@@ -92,7 +92,7 @@ class Section extends ChangeNotifier{
     }
 
     for(final original in originalItems){
-      if(!items.contains(original)){
+      if(!items.contains(original) && (original.image as String).contains('firebase')){
         try{
           final ref = await storage.getReferenceFromUrl(
             original.image as String
@@ -113,13 +113,14 @@ class Section extends ChangeNotifier{
   Future<void> delete() async{
     await firestoreRef.delete();
     for(final item in items){
-      try{
-        final ref = await storage.getReferenceFromUrl(
-          item.image as String
-        );
-        await ref.delete();
-      // ignore: empty_catches
-      } catch(e){}
+      if((item.image as String).contains('firebase'))
+        try{
+          final ref = await storage.getReferenceFromUrl(
+            item.image as String
+          );
+          await ref.delete();
+        // ignore: empty_catches
+        } catch(e){}
     }
   }
